% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.installed.R
\name{check.installed}
\alias{check.installed}
\title{checks if the required libraries are installed in the system}
\usage{
check.installed(libraries)
}
\arguments{
\item{libraries}{A vector with the libraries that the project uses}
}
\description{
checks if the required libraries are installed in the system
}
