% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gmdac.R
\docType{data}
\name{IOM_colors}
\alias{IOM_colors}
\title{Color palette for GMDAC visualization}
\format{
An object of class \code{character} of length 18.
}
\usage{
IOM_colors
}
\arguments{
\item{...}{name of the colours from the gmdac pallete}
}
\value{
colors passes to \code{colors_gmdac}
}
\description{
Color palette for GMDAC visualization
}
\examples{
# get the names of all the colors in the palette:
names(colors_gmdac())
get a color:
colors_gmdac("blue_ligth")
get two colors:
colors_gmdac(c("blue_ligth", "red"))
}
\author{
Murtiza Numamat (murtiza.nurmamat@gmail.com)
Andres Arau (jarau@iom.int)
}
\keyword{datasets}
