% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{Loads all the functions created in a project}
\usage{
load_functions(dir)
}
\arguments{
\item{dir}{A directory where the functions are saved in the project i.e. functions/}
}
\description{
Loads all the functions created in a project
}
\examples{
Define a directory where all the functions are saved (dir_functions <- functions)
Run \code{load_functions(dir_functions)} to load all the functions at the same time
}
\author{
Andres Arau (jarau@iom.int)
}
