% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size}
\alias{sample_size}
\title{Sample size calculator based on:
https://www.qualtrics.com/uk/experience-management/research/determine-sample-size/?rid=ip&prevsite=en&newsite=uk&geo=ES&geomatch=uk
also based on a simple calculation by survey monkey that assumes a s.e of .25
https://www.surveymonkey.com/mp/sample-size-calculator/}
\usage{
sample_size(N, margin_error, confidence_interval)
}
\arguments{
\item{N}{population}

\item{margin_error}{(diff between the mean number of sample and mean number of population)
has to be expressed in decimals (0.05, 0.02)}

\item{confidence}{interval of how confident you want to be that the actual mean falls within the error margin (expressed
in z-score 90% = 1.645, 95% 1.96, 99% 2.576)}
}
\description{
Sample size calculator based on:
https://www.qualtrics.com/uk/experience-management/research/determine-sample-size/?rid=ip&prevsite=en&newsite=uk&geo=ES&geomatch=uk
also based on a simple calculation by survey monkey that assumes a s.e of .25
https://www.surveymonkey.com/mp/sample-size-calculator/
}
