% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hi.R
\name{say_hi}
\alias{say_hi}
\title{Say hello to a team member}
\usage{
say_hi(name)
}
\arguments{
\item{name}{A character with your name}
}
\value{
A message saying hi!
}
\description{
Say hello to a team member
}
