% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_colors.R
\docType{data}
\name{un_pallete}
\alias{un_pallete}
\title{pallette with un colors
taken from https://www.un.org/sustainabledevelopment/}
\format{
An object of class \code{character} of length 19.
}
\usage{
un_pallete
}
\value{
a UN color
}
\description{
pallette with un colors
taken from https://www.un.org/sustainabledevelopment/
}
\keyword{datasets}
