% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_value_from_list.R
\name{unnest_value_from_list}
\alias{unnest_value_from_list}
\title{To unnest variables of class list}
\usage{
unnest_value_from_list(x)
}
\arguments{
\item{x}{Variable in a dataset that is of list class}
}
\description{
To unnest variables of class list
}
