% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoho_get_report.R
\name{zoho_get_report}
\alias{zoho_get_report}
\title{Gets report from Zoho based on criteria}
\usage{
zoho_get_report(zoho_report, new_token, criteria)
}
\arguments{
\item{zoho_report}{A string. Report link name as it is defined in Zoho}

\item{access_token}{A string. The refresh token that belongs to the requesting user,
which is returned when the access_type=offline is included in the query string of the authorization request
See \href{https://www.zoho.com/creator/help/api/v2/refresh-the-access-token.html}{Refresh the Access Token}}
}
\value{
A tibble. Tibble format of the requested report
}
\description{
Gets report from Zoho based on criteria
}
